+++ source/Nuke.Azure.KeyVault/Attributes/KeyVaultAttribute.cs
@@ -4,7 +4,10 @@
 
 using System;
 using System.Linq;
+using System.Reflection;
 using JetBrains.Annotations;
+using Nuke.Common;
+using Nuke.Common.Execution;
 
 namespace Nuke.Azure.KeyVault
 {
@@ -12,11 +15,11 @@ namespace Nuke.Azure.KeyVault
     [PublicAPI]
     public class KeyVaultAttribute : KeyVaultSecretAttribute
     {
-        public override object GetValue (string memberName, Type memberType)
+        public override object GetValue (MemberInfo member, NukeBuild build)
         {
-            if (memberType != typeof(KeyVault))
+            if (member.GetFieldOrPropertyType() != typeof(KeyVault))
                 throw new NotSupportedException();
-            return base.GetValue(memberName, memberType);
+            return base.GetValue(member, build);
         }
     }
 }