+++ source/Nuke.Azure.KeyVault/Attributes/KeyVaultKeyAttribute.cs
@@ -4,7 +4,10 @@
 
 using System;
 using System.Linq;
+using System.Reflection;
 using JetBrains.Annotations;
+using Nuke.Common;
+using Nuke.Common.Execution;
 
 namespace Nuke.Azure.KeyVault
 {
@@ -19,11 +22,11 @@ namespace Nuke.Azure.KeyVault
         {
         }
 
-        public override object GetValue ([NotNull] string memberName, [NotNull] Type memberType)
+        public override object GetValue (MemberInfo member, NukeBuild build)
         {
-            if (memberType != typeof(KeyVaultKey))
+            if (member.GetFieldOrPropertyType() != typeof(KeyVaultKey))
                 throw new NotSupportedException();
-            return base.GetValue(memberName, memberType);
+            return base.GetValue(member, build);
         }
     }
 }